+++ src/main.rs
@@ -12,7 +12,6 @@ mod labels;
 mod listeners;
 mod network;
 mod packet;
-mod recorder;
 mod telemetry;
 mod theme;
 mod ui;